+++ src/Slide.jsx
@@ -15,6 +15,8 @@ export default class Slide extends React.Component {
             React.PropTypes.node
         ]),
         className: React.PropTypes.string,
+        componentDidSlideIn: React.PropTypes.func,
+        componentWillSlideIn: React.PropTypes.func,
         direction: React.PropTypes.oneOf([ 'up', 'down' ]),
         slideInDuration: React.PropTypes.number,
         slideOutDuration: React.PropTypes.number,
@@ -55,6 +57,8 @@ export default class Slide extends React.Component {
             animateOnSlideOut,
             children,
             className,
+            componentWillSlideIn,
+            componentDidSlideIn,
             slideInDuration,
             slideOutDuration,
             style,
@@ -68,6 +72,8 @@ export default class Slide extends React.Component {
         );
 
         const animationProps = {
+            componentDidEnter: componentDidSlideIn,
+            componentWillEnter: componentWillSlideIn,
             transitionAppear: false,
             transitionEnter: animateOnSlideIn,
             transitionEnterTimeout: slideInDuration,