+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+<a name="0.2.0"></a>
+## 0.2.0 (2016-04-27)
+
+* feat(expand-animation): add life-cycle hooks for the entering component ([96bfda3](https://github.com/telerik/kendo-react-animation/commit/96bfda3))
+* feat(fade-animation): add life-cycle hooks for the entering component ([28a1143](https://github.com/telerik/kendo-react-animation/commit/28a1143))
+* feat(slide-animation): add life-cycle hooks for the entering component ([be1ee1b](https://github.com/telerik/kendo-react-animation/commit/be1ee1b))
+
 <a name="0.1.0"></a>
 # 0.1.0 (2016-04-26)